Add unit tests for transport options in config package
